miner changed

+++ src/main/java/com/example/employee/users/repository.java
@@ -5,5 +5,5 @@ import com.example.employee.users.model;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface repository extends MongoRepository<model,Long> {
+public interface repository extends MongoRepository<model,Integer> {
 }
